AT-127: add routing, login, and signup pages
