+++ frontend/src/app/app.component.html
@@ -1,22 +1,1 @@
-<app-toolbar></app-toolbar>
-<div class="columns">
-  <div class="column is-one-fifth">
-    <app-sidebar (searchChanged)="doSearch($event)"></app-sidebar>
-  </div>
-  <div class="column is-two-fifths">
-    <div class="job-list-content-container">
-      <mat-spinner *ngIf="loading" mode="indeterminate"></mat-spinner>
-      <div class="job-list-items">
-        <app-job-list *ngIf="!loading && jobs" [jobs]="jobs" (jobClicked)="onJobClicked($event)"></app-job-list>
-      </div>
-      <div class="job-list-paginator" [hidden]="loading">
-        <mat-paginator #paginator [showFirstLastButtons]="true" [length]=9999 [pageSize]="10" [pageSizeOptions]="[10, 25, 50, 100]" (page)="onPageChange($event)"></mat-paginator>
-      </div>
-    </div>
-  </div>
-  <div class="column is-two-fifths">
-    <div class="job-details-container">
-      <app-job-details *ngIf="!loading && selectedJob" [job]="selectedJob"></app-job-details>
-    </div>
-  </div>
-</div>
+<router-outlet></router-outlet>
