+++ frontend/src/app/main-layout/main-layout.component.html
@@ -1,0 +1,6 @@
+<div class="toolbar">
+  <app-toolbar></app-toolbar>
+</div>
+<div class="page-content">
+  <router-outlet></router-outlet>
+</div>
