+++ frontend/src/app/pages/login/login.component.html
@@ -1,0 +1,1 @@
+<p>login works!</p>
