+++ frontend/src/app/pages/sign-up/sign-up.component.html
@@ -1,0 +1,1 @@
+<p>sign-up works!</p>
