+++ frontend/src/app/toolbar/toolbar.component.html
@@ -1,3 +1,10 @@
-<mat-toolbar color="primary" class="toolbar" class="mat-elevation-z4">
-    <span class="agtern-title">AgTern</span>
+<mat-toolbar color="primary">
+  <span>AgTern</span>
+  <span class="toolbar-spacer"></span>
+  <div class="toolbar-bttn-container">
+    <a mat-stroked-button color="primary" routerLink="" aria-label="Home page button">Home</a>
+    <a mat-stroked-button color="primary" routerLink="jobs" aria-label="Jobs page button">Jobs</a>
+    <a mat-raised-button color="primary" routerLink="login" aria-label="Log in page button">Log In</a>
+    <a mat-button routerLink="sign-up" aria-label="Sign up page button">Sign Up</a>
+  </div>
 </mat-toolbar>
